Cleanup algebra solvers
- Moves arithmetic only solver versions to prim/functor
- Makes make_unsafe_chainable_ptr() for handling the partial pivot LU
- Reduces the memory usage in the solvers
- General cleanup
