+++ stan/math/rev/functor/algebra_solver_newton.hpp
@@ -181,14 +181,13 @@ Eigen::Matrix<var, Eigen::Dynamic, 1> algebra_solver_newton_impl(
   jacobian(f_wrt_x, theta_dbl, f_x, Jf_x);
 
   using ret_type = Eigen::Matrix<var, Eigen::Dynamic, -1>;
-  auto arena_Jf_x = to_arena(Jf_x);
-
   arena_t<ret_type> ret = theta_dbl;
+  auto Jf_xT_lu_ptr
+      = make_unsafe_chainable_ptr(Jf_x.transpose().partialPivLu());  // Lu
 
-  reverse_pass_callback([f, ret, arena_args_tuple, arena_Jf_x, msgs]() mutable {
+  reverse_pass_callback([f, ret, arena_args_tuple, Jf_xT_lu_ptr, msgs]() mutable {
     // Contract specificities with inverse Jacobian of f with respect to x.
-    Eigen::VectorXd ret_adj = ret.adj();
-    Eigen::VectorXd eta = -arena_Jf_x.transpose().lu().solve(ret_adj);
+    Eigen::VectorXd eta = -Jf_xT_lu_ptr->solve(ret.adj().eval());
 
     // Contract with Jacobian of f with respect to y using a nested reverse
     // autodiff pass.
